src/components/NewCard/NewCard.spec.tsx
import * as React from 'react';
import { createUniDriverFactory } from 'wix-ui-test-utils/uni-driver-factory';
import { isUniEnzymeTestkitExists } from 'wix-ui-test-utils/enzyme';
import { isUniTestkitExists } from 'wix-ui-test-utils/vanilla';
import { mount } from 'enzyme';
import { TPAComponentsWrapper } from '../../test/utils';
import { newCardDriverFactory } from './NewCard.driver';
import { NewCard } from './';
import { newCardTestkitFactory } from '../../testkit';
import { newCardTestkitFactory as enzymeNewCardTestkitFactory } from '../../testkit/enzyme';

describe('NewCard', () => {
  const createDriver = createUniDriverFactory(newCardDriverFactory);

  it('should render', async () => {
    const driver = createDriver(<NewCard buttonText="Click Me" />);
    expect(await driver.exists()).toBe(true);
  });

  describe('testkit', () => {
    it('should exist', async () => {
      expect(
        await isUniTestkitExists(<NewCard />, newCardTestkitFactory, {
          dataHookPropName: 'data-hook',
        }),
      ).toBe(true);
    });
  });

  describe('enzyme testkit', () => {
    it('should exist', async () => {
      expect(
        await isUniEnzymeTestkitExists(
          <NewCard />,
          enzymeNewCardTestkitFactory,
          mount,
          {
            dataHookPropName: 'data-hook',
          },
        ),
      ).toBe(true);
    });
  });
});
